Apply `--suggest-unsafe` setting to "useless cast"

+++ test/cli/autocorrect-useless-cast/test.sh
@@ -1,19 +1,50 @@
 #!/bin/bash
 
-tmp="$(mktemp -d)"
-infile="test/cli/autocorrect-useless-cast/autocorrect-useless-cast.rb"
-cp "$infile" "$tmp"
-
 cwd="$(pwd)"
-cd "$tmp" || exit 1
+infile="$cwd/test/cli/autocorrect-useless-cast/autocorrect-useless-cast.rb"
 
-"$cwd/main/sorbet" --silence-dev-message -a autocorrect-useless-cast.rb 2>&1
+tmp="$(mktemp -d)"
+
+(
+  cp "$infile" "$tmp"
+
+  cd "$tmp" || exit 1
+  if "$cwd/main/sorbet" --silence-dev-message -a autocorrect-useless-cast.rb 2>&1; then
+    echo "Expected to fail!"
+    exit 1
+  fi
+
+  echo
+  echo --------------------------------------------------------------------------
+  echo
+
+  # Also cat the file, to make that the autocorrect applied
+  cat autocorrect-useless-cast.rb
+
+  rm autocorrect-useless-cast.rb
+)
 
 echo
 echo --------------------------------------------------------------------------
 echo
 
-cat autocorrect-useless-cast.rb
+(
+  cp "$infile" "$tmp"
 
-rm autocorrect-useless-cast.rb
-rm "$tmp"
+  cd "$tmp" || exit 1
+  if "$cwd/main/sorbet" --silence-dev-message --suggest-unsafe -a autocorrect-useless-cast.rb 2>&1; then
+    echo "Expected to fail!"
+    exit 1
+  fi
+
+  echo
+  echo --------------------------------------------------------------------------
+  echo
+
+  # Also cat the file, to make that the autocorrect applied
+  cat autocorrect-useless-cast.rb
+
+  rm autocorrect-useless-cast.rb
+)
+
+rm -r "$tmp"
